codigo atomizado y comentado
