+++ src/components/averiasComponents/TableCrudAverias.vue
@@ -1,7 +1,11 @@
+/**
+ * Component for displaying and managing faults.
+ */
 <template>
   <v-app>
     <v-card>
       <v-card-title>
+        <!-- Search input field -->
         <v-text-field
           v-model="search"
           append-icon="mdi-magnify"
@@ -11,6 +15,7 @@
         ></v-text-field>
         <v-spacer></v-spacer>
         <div class="text-center">
+          <!-- Button to create a new fault -->
           <v-btn
             class="mx-2"
             fab
@@ -23,25 +28,28 @@
         </div>
       </v-card-title>
 
+      <!-- Faults table -->
       <v-data-table
         :headers="headers"
         :items="faults"
         :search="search"
       >
         <template v-slot:item.actions="{ item }">
+          <!-- Delete button for each fault -->
           <v-icon
             small
             class="mr-2"
             color="red"
-            @click=" confirmDelete(item)"
+            @click="confirmDelete(item)"
             >mdi-delete</v-icon
           >
         </template>
       </v-data-table>
 
+      <!-- Preventivs detail component -->
       <PreventivsDetail />
 
-      <!--Modal dialogo crear Preventivo -->
+      <!-- Modal dialog to create a new fault -->
       <v-dialog v-model="dialogCreateFault">
         <CreateFault @dialogCreateFault="handlePreventivCreated" />
       </v-dialog>
@@ -51,11 +59,10 @@
 </template>
 
 <script>
-import { db, collection, getDocs, deleteDoc, doc } from "@/main";
 
 import CreateFault from "@/components/averiasComponents/CreateFault.vue";
 import PreventivsDetail from "@/components/preventivsComponents/PreventivsDetail.vue";
-
+import faultRepository from "@/repository/faultRepository";
 export default {
   components: {
     CreateFault,
@@ -63,19 +70,15 @@ export default {
   },
   data() {
     return {
-      dialogDeleteFault: false,
-      dialogPreventivFault: false,
-      search: "",
-      dialogCreateFault: false,
-      machines: [],
-      selectedFault: [],
-      faults:[],
-      headers: [
-        { text: "Cod.Acceso", align: "start", filterable: false, value: "accessCode" },
-        { text: "Cod.Máquina", value: "machineCode.machineCode" },
-        { text: "Fecha Inicio", align: "center", value: "startDate" },
-        { text: "Estado", align: "center", value: "state" },
-        { text: "Acciones", align: "center", value: "actions" },
+      search: "", // Search input value
+      dialogCreateFault: false, // Dialog flag to create a new fault
+      faults: [], // List of faults
+      headers: [ // Table headers for faults
+        { text: "Access Code", align: "start", filterable: false, value: "accessCode" },
+        { text: "Machine Code", value: "machineCode.machineCode" },
+        { text: "Start Date", align: "center", value: "startDate" },
+        { text: "Status", align: "center", value: "state" },
+        { text: "Actions", align: "center", value: "actions" },
       ],
     };
   },
@@ -85,41 +88,48 @@ export default {
   },
 
   methods: {
+    /**
+     * Confirms the deletion of a fault.
+     * @param {Object} item - The fault item to delete.
+     */
     confirmDelete(item) {
-      if (confirm("¿Estás seguro de que deseas eliminar?")) {
+      if (confirm("Are you sure you want to delete?")) {
         this.deleteFault(item);
         this.getAll();
       }
     },
+    /**
+     * Handles the event when a new fault is created.
+     */
     handlePreventivCreated() {
       this.dialogCreateFault = false;
       this.getAll();
     },
+    /**
+     * Deletes a fault from the database.
+     * @param {Object} item - The fault item to delete.
+     */
     async deleteFault(item) {
+
       try {
-        const docRef = doc(db, "averias", item.id);
-        await deleteDoc(docRef);
-        this.getAll();
+        await faultRepository.delete(item)
       } catch (error) {
         console.log(error);
       }
+ 
     },
+    /**
+     * Retrieves all faults from the database.
+     */
     async getAll() {
+
       try {
-        const snapshot = await getDocs(collection(db, "averias"));
-        const faultsDb = [];
-
-        snapshot.forEach((doc) => {
-          let faultData = doc.data();
-          faultData.id = doc.id;
-          faultsDb.push(faultData);
-        });
-
-        this.faults = faultsDb;
+        
+        this.faults = await faultRepository.getAll()
       } catch (error) {
         console.log(error);
       }
-    },
+    }
   },
 };
 </script>
