+++ src/components/preventivsComponents/PreventivsDetail.vue
@@ -6,7 +6,9 @@
         <v-row>
           <v-col cols="6">
             <v-card class="mb-3">
-              <v-card-title class="subtitle-1">Información General</v-card-title>
+              <v-card-title class="subtitle-1"
+                >Información General</v-card-title
+              >
               <v-card-text>
                 <v-row>
                   <v-col cols="6">
@@ -59,7 +61,7 @@
                 <v-row>
                   <v-col cols="6">
                     <v-text-field
-                    type="date"
+                      type="date"
                       label="Fecha de inicio"
                       v-model="itemRecibido.startDate"
                       :readonly="isReadOnly"
@@ -92,7 +94,9 @@
           <v-col cols="6">
             <v-col cols="12">
               <v-card class="mb-3">
-                <v-card-title class="subtitle-1">Información de la Máquina</v-card-title>
+                <v-card-title class="subtitle-1"
+                  >Información de la Máquina</v-card-title
+                >
                 <v-card-text>
                   <v-row>
                     <v-col cols="12">
@@ -121,7 +125,6 @@
                         v-model="itemRecibido.machineCode.location"
                         :readonly="isReadOnly"
                       >
-          
                       </v-text-field>
                     </v-col>
                     <v-col cols="6">
@@ -151,7 +154,9 @@
                   {{ tarea.nameTarea | uppercase }}
                 </v-list-item-content>
               </template>
-              <v-list-item-subtitle class="text-center">Procesos de la tarea</v-list-item-subtitle>
+              <v-list-item-subtitle class="text-center"
+                >Procesos de la tarea</v-list-item-subtitle
+              >
               <v-list>
                 <v-list-item
                   v-for="(datos, repuestoIndex) in tarea.datos"
@@ -162,7 +167,9 @@
                   </v-list-item-content>
                 </v-list-item>
               </v-list>
-              <v-list-item-subtitle class="text-center">Repuestos necesarios:</v-list-item-subtitle>
+              <v-list-item-subtitle class="text-center"
+                >Repuestos necesarios:</v-list-item-subtitle
+              >
               <v-list>
                 <v-list-item
                   v-for="(repuesto, repuestoIndex) in tarea.repuestos"
@@ -179,58 +186,87 @@
             </v-list-group>
           </v-list>
         </v-card>
-        <v-btn v-if="!isReadOnly" @click="confirmUpdate" :disabled="!hasChanges">Actualizar</v-btn>
+        <v-btn v-if="!isReadOnly" @click="confirmUpdate" :disabled="!hasChanges"
+          >Actualizar</v-btn
+        >
       </v-container>
     </v-card-text>
   </v-card>
 </template>
 
 <script>
-import { updateDoc } from "firebase/firestore";
-import { db,doc } from "@/main";
-
 import eventBus from "@/config/eventBus";
+import preventivRepository from "@/repository/preventivRepository";
 
 export default {
   data() {
+    /**
+     * Vue data properties
+     * @returns {Object}
+     */
     return {
-      itemRecibido: null,
-      isReadOnly: true,
-      originalItem: null,
-      hasChanges: false
+      itemRecibido: null, // Received item
+      isReadOnly: true, // Flag indicating if the item is in read-only mode
+      originalItem: null, // Copy of the original item
+      hasChanges: false, // Flag indicating if there are any changes
     };
   },
   computed: {
+    /**
+     * Computed properties for the component
+     * @returns {Object}
+     */
     estadoClase() {
+      /**
+       * CSS class based on the state of the item
+       * @returns {String} - CSS class
+       */
       const estado = this.itemRecibido.state;
-      if (estado === "Acabado") {
+      if (estado === "Completed") {
         return "green--text"; // Class for green text
-      } else if (estado === "En proceso") {
+      } else if (estado === "In Progress") {
         return "orange--text"; // Class for orange text
-      } else if (estado === "Pendiente") {
+      } else if (estado === "Pending") {
         return "red--text"; // Class for red text
       }
       return ""; // No additional class
     },
   },
   created() {
+    /**
+     * Vue lifecycle hook: created
+     * Called when the component is created
+     */
     eventBus.$on("item-selected", this.procesarItem);
   },
   methods: {
+    /**
+     * Confirms the update of the item
+     * @returns {void}
+     */
     confirmUpdate() {
-      console.log("voy a actualizar " + this.itemRecibido.student)
-      if (confirm("¿Estás seguro de que deseas actualizar?")) {
+      if (confirm("Are you sure you want to update?")) {
         this.upDate(this.itemRecibido);
       }
     },
+    /**
+     * Processes the received item
+     * @param {Object} item - Received item
+     * @returns {void}
+     */
     procesarItem(item) {
       this.itemRecibido = item;
       this.originalItem = JSON.parse(JSON.stringify(item));
       this.isReadOnly = true;
       this.hasChanges = false;
     },
+    /**
+     * Toggles the edit mode for a property
+     * @param {String} prop - Property to toggle
+     * @returns {void}
+     */
     toggleEdit(prop) {
-      if (this.itemRecibido.state !== "Acabado") {
+      if (this.itemRecibido.state !== "Completed") {
         this.isReadOnly = !this.isReadOnly;
         if (!this.isReadOnly) {
           this.originalItem = JSON.parse(JSON.stringify(this.itemRecibido));
@@ -240,51 +276,67 @@ export default {
         this.hasChanges = false;
       }
     },
-
+    /**
+     * Updates the item
+     * @param {Object} itemRecibido - Item to update
+     * @returns {void}
+     */
     async upDate(itemRecibido) {
-      console.log("voy a actualizar")
-  try {
-    const docRef = doc(db, "preventivos", itemRecibido.id);
-    await updateDoc(docRef, {
+      try {
+        const preventData = {
+          namePersonInCharge: this.itemRecibido.namePersonInCharge,
+          accessCode: this.itemRecibido.accessCode,
+          student: this.itemRecibido.student,
+          password: this.itemRecibido.password,
+          startDate: this.itemRecibido.startDate,
+          state: this.itemRecibido.state,
+          machineCode: this.itemRecibido.machineCode,
+          tareas: this.itemRecibido.tareas,
+        };
+        await preventivRepository.upDate(preventData);
+      } catch (error) {
+        console.error(error);
+      }
 
-      namePersonInCharge: this.itemRecibido.namePersonInCharge,
-      accessCode: this.itemRecibido.accessCode,
-      student: this.itemRecibido.student,
-      password: this.itemRecibido.password,
-      startDate: this.itemRecibido.startDate,
-      state: this.itemRecibido.state,
-      machineCode: this.itemRecibido.machineCode,
-      tareas: this.itemRecibido.tareas,
-    });
-  } catch (error) {
-    console.log(error);
-  }
-  eventBus.$emit("changes-saved");
-  this.isReadOnly = true;
-  this.hasChanges = false;
-},
+      eventBus.$emit("changes-saved");
+      this.isReadOnly = true;
+      this.hasChanges = false;
+    },
+    /**
+     * Removes a task from the item
+     * @param {Number} index - Index of the task to remove
+     * @returns {void}
+     */
     eliminarTarea(index) {
       this.itemRecibido.tareas.splice(index, 1);
-    }
+    },
   },
   watch: {
     itemRecibido: {
       deep: true,
+      /**
+       * Watcher for the itemRecibido property
+       * Called when the itemRecibido property changes
+       * @returns {void}
+       */
       handler() {
-        if (JSON.stringify(this.itemRecibido) !== JSON.stringify(this.originalItem)) {
+        if (
+          JSON.stringify(this.itemRecibido) !==
+          JSON.stringify(this.originalItem)
+        ) {
           this.hasChanges = true;
         } else {
           this.hasChanges = false;
         }
-      }
-    }
-  }
+      },
+    },
+  },
 };
 </script>
 
 <style scoped>
 .custom-card {
-  border: 2px solid #ACF3DB; /* Color de borde #E1EA1F */
+  border: 2px solid #acf3db; /* Color de borde #E1EA1F */
   border-radius: 4px; /* Opcional: agrega esquinas redondeadas */
 }
 .subtitle-1 {
