+++ src/components/tareasComponents/TableCrudTask.vue
@@ -11,13 +11,12 @@
       <v-spacer></v-spacer>
       <template>
         <div class="text-center">
-        
           <v-btn
             class="mx-2"
             fab
             dark
             color="indigo"
-            @click="getRepuestos(),dialogCreatTarea = true"
+            @click="dialogCreatTarea = true"
           >
             <v-icon dark>mdi-plus</v-icon>
           </v-btn>
@@ -29,20 +28,31 @@
       :headers="headers"
       :items="tareas"
       :search="search"
-      @click:row="selectedTarea = $event; sendItem(selectedTarea)"
+      @click:row="
+        selectedTarea = $event;
+        sendItem(selectedTarea);
+      "
     >
       <template v-slot:item.actions="{ item }">
         <v-icon
           small
           class="mr-2"
           color="red"
-          @click="selectedTarea = item; dialogDelete = true"
-        >mdi-delete</v-icon>
+          @click="
+            selectedTarea = item;
+            confirmDelete(selectedTarea);
+          "
+          >mdi-delete</v-icon
+        >
         <v-icon
           small
           color="blue"
-          @click="selectedTarea = item; dialogUpDateTarea = true"
-        >mdi-pencil</v-icon>
+          @click="
+            selectedTarea = item;
+            dialogUpDateTarea = true;
+          "
+          >mdi-pencil</v-icon
+        >
       </template>
     </v-data-table>
 
@@ -91,7 +101,8 @@
                       color="primary"
                       @click="agregarDato"
                       :disabled="nuevoDato === ''"
-                    >Agregar</v-btn>
+                      >Agregar</v-btn
+                    >
                   </v-col>
                 </v-row>
               </v-col>
@@ -111,27 +122,60 @@
                     <template v-slot:item="{ item }">
                       <tr>
                         <td>{{ item }}</td>
+                        <v-icon
+                          small
+                          class="mr-2"
+                          color="red"
+                          @click="deleteDatos(item)"
+                        >mdi-delete</v-icon>
                       </tr>
                     </template>
                   </v-data-table>
                 </v-card>
               </v-col>
             </v-row>
+            <v-row>
+          <v-col cols="12">
+            <v-card>
+              <h4>Lista de repuestos necesarios</h4>
+    
+              <v-data-table
+                v-model="selectedTarea.selectedRepuesto"
+                :headers="headersTablaRepuestos"
+                :items="selectedTarea.repuestos"
+                :single-select="singleSelect"
+                item-key="id"
+                show-select
+                class="elevation-1"
+                @input="updateSelectedItems"
+              >
+                <template v-slot:top>
+                  <v-toolbar flat>
+                    <v-toolbar-title>Repuestos</v-toolbar-title>
+                    <v-spacer></v-spacer>
+                    <v-text-field
+                      v-model="search"
+                      append-icon="mdi-magnify"
+                      label="Buscar"
+                      single-line
+                      hide-details
+                    ></v-text-field>
+                  </v-toolbar>
+                </template>
+              </v-data-table>
+            </v-card>
+          </v-col>
 
-            <v-card-actions>
-              <v-btn color="primary" text @click="dialogUpDateTarea = false">
-                Cancelar
-              </v-btn>
-              <v-spacer></v-spacer>
-              <v-btn color="red" text type="submit">
-                Eliminar
-              </v-btn>
-              <v-btn
-                color="green"
-                text
-                @click="updateTarea(selectedTarea)"
-              >Actualizar</v-btn>
-            </v-card-actions>
+            </v-row>
+
+            <v-btn
+              color="primary"
+              class="mr-4"
+              text
+              @click="confirmUpDate(selectedTarea); dialogUpDateTarea = false"
+              :disabled="!hayCambios"
+              >Actualizar</v-btn
+            >
           </v-form>
         </v-container>
       </v-card>
@@ -139,162 +183,29 @@
 
     <!-- Modal Crear tarea -->
     <v-dialog v-model="dialogCreatTarea">
-      <v-card>
-        <v-container>
-          <v-form @submit.prevent="addTarea">
-            <p><b>AGREGAR TAREA</b></p>
-
-            <v-row>
-              <v-col cols="6">
-                <v-text-field
-                  type="text"
-                  label="Agregar nombre de tarea"
-                  v-model="nameTarea"
-                ></v-text-field>
-              </v-col>
-              <v-col cols="6">
-                <v-text-field
-                  type="text"
-                  label="Categoria"
-                  v-model="category"
-                ></v-text-field>
-              </v-col>
-            </v-row>
-            <v-row>
-              <v-col cols="6">
-                <v-combobox
-                  v-model="selectedFrencunce"
-                  :items="frecuencias"
-                  label="Selecciona una frecuencia"
-                ></v-combobox>
-              </v-col>
-              <v-col cols="6">
-                <v-row>
-                  <v-col cols="9">
-                    <v-text-field
-                      type="text"
-                      label="Describe una operación"
-                      v-model="nuevoDato"
-                    ></v-text-field>
-                  </v-col>
-                  <v-col cols="3" class="align-self-center">
-                    <v-btn
-                      color="primary"
-                      @click="agregarDato"
-                      :disabled="nuevoDato === ''"
-                    >Agregar</v-btn>
-                  </v-col>
-                </v-row>
-              </v-col>
-            </v-row>
-            <v-row>
-              <v-col cols="12">
-                <v-card>
-                  <v-card-title>
-                    <h5>Pasos de tarea</h5>
-                  </v-card-title>
-
-                  <v-data-table
-                    :items="datos"
-                    hide-default-footer
-                    class="elevation-1"
-                  >
-                    <template v-slot:item="{ item }">
-                      <tr>
-                        <td>{{ item }}</td>
-                      </tr>
-                    </template>
-                  </v-data-table>
-                </v-card>
-              </v-col>
-            </v-row>
-            <v-row>
-              <v-col cols="12">
-                <v-card>
-                  <h4>Lista de repuestos necesarios</h4>
-                  <v-text-field
-        v-model="searchRepuestos"
-        append-icon="mdi-magnify"
-        label="Buscar"
-        single-line
-        hide-details
-      ></v-text-field>
-      <v-data-table
-    v-model="selectedRepuesto"
-    :headers="headersTablaRepuestos"
-    :items="repuestosDB"
-    :single-select="singleSelect"
-    item-key="id"
-    show-select
-    class="elevation-1"
-    @input="updateSelectedItems"
-  >
-    <template v-slot:top>
-      <v-toolbar flat>
-        <v-toolbar-title>Repuestos</v-toolbar-title>
-        <v-spacer></v-spacer>
-        <v-text-field
-          v-model="search"
-          append-icon="mdi-magnify"
-          label="Buscar"
-          single-line
-          hide-details
-        ></v-text-field>
-      </v-toolbar>
-    </template>
-  </v-data-table>
-                </v-card>              
-              </v-col>
-            </v-row>
-              <br/>
-              <v-btn  color="primary"
-                class="mr-4"  type="submit">
-                Agregar Tarea
-              </v-btn>
-          </v-form>
-        </v-container>
-      </v-card>
-    </v-dialog>
-
-    <!-- Modal Confirmar eliminación de tarea -->
-    <v-dialog v-model="dialogDelete">
-      <v-card>
-        <v-card-title class="headline">Confirmar eliminación</v-card-title>
-        <v-card-text>¿Estás seguro de que deseas eliminar esta tarea?</v-card-text>
-        <v-card-actions>
-          <v-spacer></v-spacer>
-          <v-btn color="primary" text @click="dialogDelete = false">Cancelar</v-btn>
-          <v-btn color="red" text @click="deleteTarea(selectedTarea)">Eliminar</v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-dialog>
-
-    <!-- Modal Confirmar eliminación de tarea en el formulario de actualización -->
-    <v-dialog v-model="dialogoConfirmDelete">
-      <v-card>
-        <v-card-title class="headline">Confirmar eliminación</v-card-title>
-        <v-card-text>¿Estás seguro de que deseas eliminar esta tarea?</v-card-text>
-        <v-card-actions>
-          <v-spacer></v-spacer>
-          <v-btn color="primary" text @click="dialogoConfirmDelete = false">Cancelar</v-btn>
-          <v-btn color="red" text @click="deleteTarea(selectedTarea)">Eliminar</v-btn>
-        </v-card-actions>
-      </v-card>
+      <CreateTask />
     </v-dialog>
   </v-card>
 </template>
 
 <script>
-import { updateDoc } from "firebase/firestore";
-import { db, collection, getDocs, addDoc, deleteDoc, doc } from "../main";
+import {
+  db,
+  collection,
+  getDocs,
+  addDoc,
+  deleteDoc,
+  doc,
+  updateDoc,
+} from "@/repository/dataBase";
 import eventBus from "@/config/eventBus";
-
+import CreateTask from "@/components/tareasComponents/CreateTask.vue";
 export default {
   data() {
     return {
       singleSelect: false,
       repuestos: [],
-      repuestosDB:[],
+      repuestosDB: [],
       searchRepuestos: "",
       selectedRepuesto: [],
       search: "",
@@ -302,13 +213,13 @@ export default {
       machines: [],
       dialogUpDateTarea: false,
       dialogoConfirmDelete: false,
-      
+
       typeErrors: [],
       dialogDelete: false,
       datos: [],
       item: null,
       nuevoDato: "",
-      selectedTarea: [],
+      selectedTarea: {},
       tareas: [],
       operaciones: null,
       nameTarea: "",
@@ -320,7 +231,6 @@ export default {
 
       selected: [],
       repuestos: [],
-      search: "",
       headersTablaRepuestos: [
         {
           text: "Identificador",
@@ -349,6 +259,11 @@ export default {
     this.getTareas();
   },
   computed: {
+    isFormValid() {
+      return (
+        this.nameTarea !== "" && this.category !== "" && this.nuevoDato !== ""
+      );
+    },
     filteredRepuestos() {
       return this.repuestos.filter((repuesto) => {
         const searchTerm = this.search.toLowerCase();
@@ -359,10 +274,29 @@ export default {
       });
     },
   },
+  components: {
+    CreateTask,
+  },
   methods: {
-    mostrarMensaje(){
-      console.log("El boton funciona bien ");
+    deleteDatos(item) {
+    const index = this.selectedTarea.datos.indexOf(item);
+    if (index !== -1) {
+      this.selectedTarea.datos.splice(index, 1);
+    }
+  },
+    confirmDelete(item) {
+      if (confirm("¿Estás seguro de que deseas eliminar?")) {
+        this.deleteTarea(item);
+        this.getAll();
+      }
     },
+    confirmUpDate(item) {
+      if (confirm("¿Estás seguro de que deseas modificar?")) {
+        this.updateTare(item);
+        this.getAll();
+      }
+    },
+
     mostrarDatosRepuestos() {
       const nombresSeleccionados = [];
       this.selectedRepuesto.forEach((item) => {
@@ -372,7 +306,9 @@ export default {
     },
 
     updateSelectedItems() {
-      this.selectedRepuestos = this.repuestosDB.filter(item => item.isSelected);
+      this.selectedRepuestos = this.repuestosDB.filter(
+        (item) => item.isSelected
+      );
     },
     sendItem(item) {
       eventBus.$emit("item-selected", item);
@@ -411,24 +347,7 @@ export default {
         console.log(error);
       }
     },
-    async getRepuestos() {
-      console.log("Voy a traerme los repuestos")
-      try {
-        const snapshot = await getDocs(collection(db, "repuestos"));
-        const repuestos = [];
 
-        snapshot.forEach((doc) => {
-          let repuestosData = doc.data();
-          repuestosData.id = doc.id;
-          repuestos.push(repuestosData);
-        });
-
-        this.repuestosDB = repuestos;
-        console.log("Voy a traerme los repuestos" + this.repuestosDB.length)
-      } catch (error) {
-        console.log(error);
-      }
-    },
     async getTareas() {
       try {
         const snapshot = await getDocs(collection(db, "tareas"));
@@ -437,7 +356,7 @@ export default {
         snapshot.forEach((doc) => {
           let tareasData = doc.data();
           tareasData.id = doc.id;
-          tareasDb.push(tareasData); 
+          tareasDb.push(tareasData);
         });
 
         this.tareas = tareasDb;
@@ -445,24 +364,6 @@ export default {
         console.log(error);
       }
     },
-    async addTarea() {
-      try {
-        
-          await addDoc(collection(db, "tareas"), {
-            nameTarea: this.nameTarea,
-            category: this.category,
-            selectedFrencunce: this.selectedFrencunce,
-            datos: this.datos,
-            repuestos:this.selectedRepuesto,
-            state:"Pendiente"
-          });
-          this.getTareas();
-      } catch (error) {
-        console.log(error);
-      }
-    },
   },
 };
-
-
 </script>
