+++ src/components/tareasComponents/TareaDetail.vue
