+++ src/router/index.js
@@ -1,46 +1,77 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-import Home from '../views/Home.vue'
+import Vue from 'vue';
+import VueRouter from 'vue-router';
+import Home from '../views/Home.vue';
+import { getAuth } from 'firebase/auth';
 
-Vue.use(VueRouter)
+Vue.use(VueRouter);
+
+const auth = getAuth();
 
 const routes = [
   {
     path: '/',
+    name: 'login',
+    component: () => import('../views/Login')
+  },
+  {
+    path: '/home',
     name: 'Home',
-    component: Home
+    component: Home,
+    meta: { requiresAuth: true }
   },
   {
     path: '/machines',
     name: 'machines',
-    component: () => import('../views/Machines.vue')
+    component: () => import('../views/Machines.vue'),
+    meta: { requiresAuth: true }
   },
   {
     path: '/repuestos',
     name: 'repuestos',
-    component: () => import('../views/Repuestos.vue')
+    component: () => import('../views/Repuestos.vue'),
+    meta: { requiresAuth: true }
   },
   {
     path: '/tareas',
     name: 'tareas',
-    component: () => import('../views/Tareas')
+    component: () => import('../views/Tareas'),
+    meta: { requiresAuth: true }
   },
   {
     path: '/preventivos',
     name: 'preventivos',
-    component: () => import('../views/Preventivos')
-  },  {
+    component: () => import('../views/Preventivos'),
+    meta: { requiresAuth: true }
+  },  
+  {
     path: '/averias',
     name: 'prevaveriasentivos',
-    component: () => import('../views/Averias')
+    component: () => import('../views/Averias'),
+    meta: { requiresAuth: true }
+  },
+  {
+    path: '/maquina',
+    name: 'maquina',
+    component: () => import('../views/Machine'),
+    meta: { requiresAuth: true }
   }
-
-]
+];
 
 const router = new VueRouter({
   mode: 'history',
   base: process.env.BASE_URL,
   routes
-})
+});
 
-export default router
+router.beforeEach((to, from, next) => {
+  const currentUser = auth.currentUser;
+  const requiresAuth = to.matched.some(record => record.meta.requiresAuth);
+
+  if (requiresAuth && !currentUser) {
+    next('/');
+  } else {
+    next();
+  }
+});
+
+export default router;
