+++ src/views/Machines.vue
@@ -7,7 +7,6 @@
     <Navbar />
     <v-container>
       <v-row>
-        <v-col cols="12" sm="9">
           <v-row>
             <v-col cols="12" sm="12">
               <Contact />
@@ -16,10 +15,6 @@
               <TableCrudMachine />
             </v-col>
           </v-row>
-        </v-col>
-        <v-col cols="12" sm="3">
-          <Doctor />
-        </v-col>
       </v-row>
     </v-container>
   </v-app>
@@ -31,7 +26,7 @@ import SideBar from "@/components/SideBar.vue";
 import Navbar from "../components/Navbar.vue";
 import Contact from "../components/Contact.vue";
 import Doctor from "../components/DataPicker.vue";
-import TableCrudMachine from "../components/TableCrudMachines.vue";
+import TableCrudMachine from "../components/machineComponents/TableCrudMachines.vue";
 
 export default {
   data: () => ({}),
